traffic_imc/pyproject.toml
[project]
name = "traffic_imc"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Seungbin Yim", email = "ysb06@hotmail.com"},
]
dependencies = [
    "wandb>=0.21.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "plotly>=5.0.0",
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "scikit-learn>=1.0.0",
    "scipy>=1.7.0",
    "pyyaml>=6.0.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
post_install = "python scripts/post_update.py"
copy-stgcn = "python scripts/post_update.py"

[tool.pdm.dev-dependencies]
dev = [
    "-e file:///${PROJECT_ROOT}/../songdo_metr#egg=songdo-metr",
]
